removed json and replaced by np.write / np.read
