+++ energy_demand/scripts/s_ss_raw_shapes.py
@@ -6,8 +6,8 @@ import csv
 from datetime import date
 import numpy as np
 from energy_demand.read_write import read_data
+from energy_demand.read_write import write_data
 from energy_demand.basic import date_prop
-from energy_demand.scripts import s_shared_functions
 from energy_demand.profiles import load_profile
 
 def dict_init_carbon_trust():
@@ -336,7 +336,7 @@ def run(paths, local_paths, lookups):
             # Write shapes to txt
             joint_string_name = str(sector) + "__" + str(enduse)
 
-            s_shared_functions.create_txt_shapes(
+            write_data.create_txt_shapes(
                 joint_string_name,
                 local_paths['ss_load_profiles'],
                 load_peak_shape_dh,
